+++ changelog.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added priority and scheduler whitelist to triggers.
 - Added search and status filter to queue element list.
 - Added overview of queue element.
+- Added option to create new queue element.
+- Added editing of queue elements.
+- Added option to delete queue element.
 
 ### Changed